Sierpiński triangle fractal
This algorithm draws Sierpiński triangle fractal using turtle module having as the first landmark a randomly chosen point from the interior of a triangle. The next point chosen is halfway between the point and one of the ends of the triangle, likewise, randomly chosen. And so on. At a large number of constructed points the Sierpiński fractal will result